[Refactor] cacheEvent Refactor
